add changes

+++ src/controllers/tweet.controller.js
@@ -44,6 +44,15 @@ const updateTweet = asyncHandler(async (req, res) => {
   if (!isValidObjectId(tweetId)) {
     throw new ApiError(400, "Invalid tweetId");
   }
+  const tweet = await Tweet.findById(tweetId);
+  if (!tweet) {
+    throw new ApiError(404, "Tweet not found");
+  }
+
+  if (tweet.owner.toString() !== req.user._id.toString()) {
+    throw new ApiError(403, "You are not authorized to perform this action");
+  }
+
   const { content } = req.body;
   const updateUserTweet = await Tweet.findByIdAndUpdate(
     tweetId,
@@ -69,8 +78,17 @@ const deleteTweet = asyncHandler(async (req, res) => {
   if (!isValidObjectId(tweetId)) {
     throw new ApiError(400, "Invalid tweetId");
   }
+
+  const tweet = await Tweet.findById(tweetId);
+  if (!tweet) {
+    throw new ApiError(404, "Tweet not found");
+  }
+
+  if (tweet.owner.toString() !== req.user._id.toString()) {
+    throw new ApiError(403, "You are not authorized to perform this action");
+  }
   const deleteUserTweet = await Tweet.findByIdAndDelete(tweetId, { new: true });
-  console.log(deleteUserTweet);
+
   if (!deleteUserTweet) {
     throw new ApiError(400, "This tweet is already Deleted");
   }
